move storage and methods into own package
