+++ storage/postgres.go
@@ -1,76 +1,82 @@
-package main
+package storage
 
 import (
 	"database/sql"
 	"fmt"
+
 	"log"
 	"math/rand"
 	"strings"
 	"time"
 
-	_ "github.com/mattn/go-sqlite3"
+	_ "github.com/lib/pq"
 	c "github.com/na50r/wombo-combo-go-be/constants"
 	dto "github.com/na50r/wombo-combo-go-be/dto"
 	u "github.com/na50r/wombo-combo-go-be/utility"
+
 )
 
-type SQLiteStore struct {
+type PostgresStore struct {
 	db *sql.DB
 }
 
-func NewSQLiteStore(name string) (*SQLiteStore, error) {
-	db, err := sql.Open("sqlite3", fmt.Sprintf("./%s.db", name))
-	if err != nil {
-		fmt.Println(err)
-		return nil, err
-	}
-
-	_, err = db.Exec("PRAGMA journal_mode = WAL")
-	if err != nil {
-		return nil, err
-	}
-	// Release lock in 5 seconds
-	// Reference: https://stackoverflow.com/questions/66909180/increase-the-lock-timeout-with-sqlite-and-what-is-the-default-values
-	_, err = db.Exec("PRAGMA busy_timeout = 5000")
+func NewPostgresStore(connString string) (*PostgresStore, error) {
+	//To avoid conflicts with dockerized postgres, make sure to create it with:
+	//Local
+	//docker run --name wc-postgres -e POSTGRES_PASSWORD=wc-local -p 5433:5432 -d postgres
+	//Map port 5432 of the container to 5433 of the host
+	//Adjust accordingly for deployment
+	db, err := sql.Open("postgres", connString)
 	if err != nil {
 		return nil, err
 	}
-
-	fmt.Println("Connected to the SQLite database successfully.")
-	return &SQLiteStore{db: db}, nil
+	if err := db.Ping(); err != nil {
+		return nil, err
+	}
+	fmt.Println("Connected to the Postgres database successfully.")
+	return &PostgresStore{db: db}, nil
 }
 
-func (s *SQLiteStore) createUnlockedTable() error {
+func (s *PostgresStore) createAchievementImageTable() error {
+	query := `create table if not exists achievement_image (
+		name varchar(100) primary key,
+		data bytea
+		)`
+	_, err := s.db.Exec(query)
+	return err
+}
+
+func (s *PostgresStore) createAchievementTable() error {
+	query := `create table if not exists achievement (
+		id serial primary key,
+		title varchar(100),
+		type text,
+		value text,
+		description text,
+		image_name varchar(100)
+		)`
+	_, err := s.db.Exec(query)
+	return err
+}
+
+func (s *PostgresStore) createUnlockedTable() error {
 	query := `create table if not exists unlocked (
-		username text,
-		achievement_title text,
+		username varchar(100),
+		achievement_title varchar(100),
 		primary key (username, achievement_title)
 		)`
 	_, err := s.db.Exec(query)
 	return err
 }
 
-func (s *SQLiteStore) createAchievementTable() error {
-	query := `create table if not exists achievement (
-		id integer primary key,
-		title text,
-		type text,
-		value text,
-		description text,
-		image_name text
-		)`
-	_, err := s.db.Exec(query)
-	return err
-}
-
-func (s *SQLiteStore) createAccountTable() error {
+func (s *PostgresStore) createAccountTable() error {
 	query := `create table if not exists account (
-		username text primary key,
-		image_name text,
-		password text,
+		username varchar(100) primary key,
+		image_name varchar(100),
+		password varchar(100),
 		wins integer,
 		losses integer,
-		created_at datetime,
+		created_at timestamp,
 		status text,
 		is_owner boolean,
 		new_word_count integer,
@@ -80,32 +86,23 @@ func (s *SQLiteStore) createAccountTable() error {
 	return err
 }
 
-func (s *SQLiteStore) createImageTable() error {
+func (s *PostgresStore) createImageTable() error {
 	query := `create table if not exists image (
-		name text primary key,
-		data blob
+		name varchar(100) primary key,
+		data bytea
 		)`
 	_, err := s.db.Exec(query)
 	return err
 }
 
-func (s *SQLiteStore) createAchievementImageTable() error {
-	query := `create table if not exists achievement_image (
-		name text primary key,
-		data blob
-		)`
-	_, err := s.db.Exec(query)
-	return err
-}
-
-func (s *SQLiteStore) createPlayerTable() error {
+func (s *PostgresStore) createPlayerTable() error {
 	query := `create table if not exists player (
-		name text,
-		lobby_code text,
-		image_name,
+		name varchar(100),
+		lobby_code varchar(100),
+		image_name varchar(100),
 		is_owner boolean,
 		has_account boolean,
-		target_word text,
+		target_word varchar(100),
 		points integer,
 		word_count integer,
 		new_word_count integer,
@@ -115,11 +112,11 @@ func (s *SQLiteStore) createPlayerTable() error {
 	return err
 }
 
-func (s *SQLiteStore) createLobbyTable() error {
+func (s *PostgresStore) createLobbyTable() error {
 	query := `create table if not exists lobby (
-		name text,
-		image_name text,
-		lobby_code text,
+		name varchar(100),
+		image_name varchar(100),
+		lobby_code varchar(100),
 		game_mode text,
 		player_count integer,
 		primary key (lobby_code)
@@ -127,11 +124,12 @@ func (s *SQLiteStore) createLobbyTable() error {
 	_, err := s.db.Exec(query)
 	return err
 }
-func (s *SQLiteStore) createCombinationTable() error {
+
+func (s *PostgresStore) createCombinationTable() error {
 	query := `create table if not exists combination (
-		a text,
-		b text,
-		result text,
+		a varchar(100),
+		b varchar(100),
+		result varchar(100),
 		depth integer,
 		primary key (a, b)
 		)`
@@ -139,9 +137,9 @@ func (s *SQLiteStore) createCombinationTable() error {
 	return err
 }
 
-func (s *SQLiteStore) createWordTable() error {
+func (s *PostgresStore) createWordTable() error {
 	query := `create table if not exists word (
-		word text primary key,
+		word varchar(100) primary key,
 		depth integer,
 		reachability float
 		)`
@@ -149,51 +147,38 @@ func (s *SQLiteStore) createWordTable() error {
 	return err
 }
 
-func (s *SQLiteStore) createPlayerWordTable() error {
+func (s *PostgresStore) createPlayerWordTable() error {
 	query := `create table if not exists player_word (
-		player_name text,
-		word text,
-		lobby_code text,
-		timestamp datetime default current_timestamp,
+		player_name varchar(100),
+		word varchar(100),
+		lobby_code varchar(100),
+		timestamp timestamp default current_timestamp,
 		primary key (player_name, word, lobby_code)
 		)`
 	_, err := s.db.Exec(query)
 	return err
 }
 
-func (s *SQLiteStore) createDailyWordTable() error {
+func (s *PostgresStore) createDailyWordTable() error {
 	query := `create table if not exists daily_word (
-		timestamp datetime default current_timestamp,
-		word text
+		timestamp timestamp default current_timestamp,
+		word varchar(100)
 		)`
 	_, err := s.db.Exec(query)
 	return err
 }
 
-func (s *SQLiteStore) createDailyChallengeTable() error {
+func (s *PostgresStore) createDailyChallengeTable() error {
 	query := `create table if not exists daily_challenge (
-		timestamp datetime default current_timestamp,
+		timestamp timestamp default current_timestamp,
 		word_count integer,
-		username text
+		username varchar(100)
 		)`
 	_, err := s.db.Exec(query)
 	return err
 }
 
-func (s *SQLiteStore) createSessionTable() error {
-	query := `create table if not exists session (
-		id text primary key,
-		refresh_token text,
-		username text,
-		is_revoked boolean,
-		created_at datetime,
-		expires_at datetime
-		)`
-	_, err := s.db.Exec(query)
-	return err
-}
-
-func (s *SQLiteStore) Init() error {
+func (s *PostgresStore) Init() error {
 	if err := s.createAccountTable(); err != nil {
 		return err
 	}
@@ -222,9 +207,6 @@ func (s *SQLiteStore) Init() error {
 	if err := s.createDailyChallengeTable(); err != nil {
 		return err
 	}
-	if err := s.createSessionTable(); err != nil {
-		return err
-	}
 	if err := s.createAchievementTable(); err != nil {
 		return err
 	}
@@ -237,8 +219,8 @@ func (s *SQLiteStore) Init() error {
 	return nil
 }
 
-func (s *SQLiteStore) UnlockAchievement(username, achievementTitle string) (bool, error) {
-	res, err := s.db.Exec("insert or ignore into unlocked (username, achievement_title) values (?, ?)", username, achievementTitle)
+func (s *PostgresStore) UnlockAchievement(username, achievementTitle string) (bool, error) {
+	res, err := s.db.Exec("insert into unlocked (username, achievement_title) values ($1, $2) on conflict do nothing", username, achievementTitle)
 	if err != nil {
 		return false, err
 	}
@@ -249,9 +231,9 @@ func (s *SQLiteStore) UnlockAchievement(username, achievementTitle string) (bool
 	return affected > 0, nil
 }
 
-func (s *SQLiteStore) AddAchievement(entry *AchievementEntry) error {
+func (s *PostgresStore) AddAchievement(entry *AchievementEntry) error {
 	_, err := s.db.Exec(
-		"insert into achievement (type, title, value, description, image_name) values (?, ?, ?, ?, ?)",
+		"insert into achievement (type, title, value, description, image_name) values ($1, $2, $3, $4, $5)",
 		entry.Type,
 		entry.Title,
 		entry.Value,
@@ -261,35 +243,35 @@ func (s *SQLiteStore) AddAchievement(entry *AchievementEntry) error {
 	return err
 }
 
-func (s *SQLiteStore) AddDailyChallengeEntry(wordCount int, username string) error {
+func (s *PostgresStore) AddDailyChallengeEntry(wordCount int, username string) error {
 	today := time.Now().Format("2006-01-02")
 	var oldCount int
-	err := s.db.QueryRow("select word_count from daily_challenge where username = ? and timestamp = ?", username, today).Scan(&oldCount)
+	err := s.db.QueryRow("select word_count from daily_challenge where username = $1 and timestamp = $2", username, today).Scan(&oldCount)
 	if err != nil && err != sql.ErrNoRows {
 		return err
 	}
 	if err == sql.ErrNoRows {
-		_, err = s.db.Exec("insert into daily_challenge (word_count, username, timestamp) values (?, ?, ?)", wordCount, username, today)
+		_, err = s.db.Exec("insert into daily_challenge (word_count, username, timestamp) values ($1, $2, $3)", wordCount, username, today)
 		return err
 	}
 	if oldCount > wordCount {
-		_, err = s.db.Exec("update daily_challenge set word_count = ? where username = ? and timestamp = ?", wordCount, username, today)
+		_, err = s.db.Exec("update daily_challenge set word_count = $1 where username = $2 and timestamp = $3", wordCount, username, today)
 		return err
 	}
 	return nil
 }
 
-func (s *SQLiteStore) CreateOrGetDailyWord(minReachability, maxReachability float64, maxDepth int) (string, error) {
+func (s *PostgresStore) CreateOrGetDailyWord(minReachability, maxReachability float64, maxDepth int) (string, error) {
 	log.Println("Creating or getting daily word")
 	today := time.Now().Format("2006-01-02")
 	var word string
-	err := s.db.QueryRow("select word from daily_word where timestamp = ?", today).Scan(&word)
+	err := s.db.QueryRow("select word from daily_word where timestamp = $1", today).Scan(&word)
 	if err == sql.ErrNoRows {
 		word, err := s.GetTargetWord(minReachability, maxReachability, maxDepth)
 		if err != nil {
 			return "", err
 		}
-		_, err = s.db.Exec("insert into daily_word (timestamp, word) values (?, ?)", today, word)
+		_, err = s.db.Exec("insert into daily_word (timestamp, word) values ($1, $2)", today, word)
 		if err != nil {
 			return "", err
 		}
@@ -298,10 +280,10 @@ func (s *SQLiteStore) CreateOrGetDailyWord(minReachability, maxReachability floa
 	return word, nil
 }
 
-func (s *SQLiteStore) CreateAccount(acc *Account) error {
+func (s *PostgresStore) CreateAccount(acc *Account) error {
 	query := `insert into account 
 	(username, image_name, password, wins, losses, created_at, status, is_owner, new_word_count, word_count)
-	values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
+	values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)`
 	_, err := s.db.Exec(
 		query,
 		acc.Username,
@@ -321,10 +303,10 @@ func (s *SQLiteStore) CreateAccount(acc *Account) error {
 	return nil
 }
 
-func (s *SQLiteStore) CreatePlayer(player *Player) error {
+func (s *PostgresStore) CreatePlayer(player *Player) error {
 	query := `insert into player 
 	(name, lobby_code, image_name, is_owner, has_account, target_word, points, word_count, new_word_count)
-	values (?, ?, ?, ?, ?, ?, ?, ?, ?)`
+	values ($1, $2, $3, $4, $5, $6, $7, $8, $9)`
 	_, err := s.db.Exec(
 		query,
 		player.Name,
@@ -343,10 +325,10 @@ func (s *SQLiteStore) CreatePlayer(player *Player) error {
 	return nil
 }
 
-func (s *SQLiteStore) CreateLobby(lobby *Lobby) error {
+func (s *PostgresStore) CreateLobby(lobby *Lobby) error {
 	query := `insert into lobby 
 	(name, image_name, lobby_code, game_mode, player_count)
-	values (?, ?, ?, ?, ?)`
+	values ($1, $2, $3, $4, $5)`
 	_, err := s.db.Exec(
 		query,
 		lobby.Name,
@@ -361,8 +343,8 @@ func (s *SQLiteStore) CreateLobby(lobby *Lobby) error {
 	return nil
 }
 
-func (s *SQLiteStore) GetPlayerByLobbyCodeAndName(name, lobbyCode string) (*Player, error) {
-	rows, err := s.db.Query("select * from player where name = ? and lobby_code = ?", name, lobbyCode)
+func (s *PostgresStore) GetPlayerByLobbyCodeAndName(name, lobbyCode string) (*Player, error) {
+	rows, err := s.db.Query("select * from player where name = $1 and lobby_code = $2", name, lobbyCode)
 	if err != nil {
 		return nil, err
 	}
@@ -373,13 +355,13 @@ func (s *SQLiteStore) GetPlayerByLobbyCodeAndName(name, lobbyCode string) (*Play
 	return nil, fmt.Errorf("player %s not found", name)
 }
 
-func (s *SQLiteStore) DeletePlayer(name, lobbyCode string) error {
-	_, err := s.db.Exec("delete from player where name = ? and lobby_code = ?", name, lobbyCode)
+func (s *PostgresStore) DeletePlayer(name, lobbyCode string) error {
+	_, err := s.db.Exec("delete from player where name = $1 and lobby_code = $2", name, lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) GetPlayersByLobbyCode(lobbyCode string) ([]*Player, error) {
-	rows, err := s.db.Query("select * from player where lobby_code = ?", lobbyCode)
+func (s *PostgresStore) GetPlayersByLobbyCode(lobbyCode string) ([]*Player, error) {
+	rows, err := s.db.Query("select * from player where lobby_code = $1", lobbyCode)
 	if err != nil {
 		return nil, err
 	}
@@ -395,8 +377,8 @@ func (s *SQLiteStore) GetPlayersByLobbyCode(lobbyCode string) ([]*Player, error)
 	return players, nil
 }
 
-func (s *SQLiteStore) GetAccountByUsername(username string) (*Account, error) {
-	rows, err := s.db.Query("select * from account where username = ?", username)
+func (s *PostgresStore) GetAccountByUsername(username string) (*Account, error) {
+	rows, err := s.db.Query("select * from account where username = $1", username)
 	if err != nil {
 		return nil, err
 	}
@@ -408,17 +390,17 @@ func (s *SQLiteStore) GetAccountByUsername(username string) (*Account, error) {
 	return nil, fmt.Errorf("account %s not found", username)
 }
 
-func (s *SQLiteStore) UpdateAccount(acc *Account) error {
+func (s *PostgresStore) UpdateAccount(acc *Account) error {
 	query := `update account set
-	username = ?,
-	image_name = ?,
-	password = ?,
-	wins = ?,
-	losses = ?,
-	status = ?,
-	new_word_count = ?,
-	word_count = ?
-	where username = ?`
+	username = $1,
+	image_name = $2,
+	password = $3,
+	wins = 	$4,
+	losses = $5,
+	status = $6,
+	new_word_count = $7,
+	word_count = $8
+	where username = $9`
 	_, err := s.db.Exec(
 		query,
 		acc.Username,
@@ -434,17 +416,17 @@ func (s *SQLiteStore) UpdateAccount(acc *Account) error {
 	return err
 }
 
-func (s *SQLiteStore) AddImage(data []byte, name string) error {
+func (s *PostgresStore) AddImage(data []byte, name string) error {
 	_, err := s.db.Exec(
-		"insert or replace into image (name, data) values (?, ?)",
+		"insert into image (name, data) values ($1, $2) on conflict (name) do update set data = $2",
 		name,
 		data,
 	)
 	return err
 }
 
-func (s *SQLiteStore) GetImage(name string) ([]byte, error) {
-	rows, err := s.db.Query("select * from image where name = ?", name)
+func (s *PostgresStore) GetImage(name string) ([]byte, error) {
+	rows, err := s.db.Query("select * from image where name = $1", name)
 	if err != nil {
 		return nil, err
 	}
@@ -466,7 +448,7 @@ func (s *SQLiteStore) GetImage(name string) ([]byte, error) {
 	return images[0].Data, nil
 }
 
-func (s *SQLiteStore) GetImages() ([]*Image, error) {
+func (s *PostgresStore) GetImages() ([]*Image, error) {
 	rows, err := s.db.Query("select * from image")
 	if err != nil {
 		return nil, err
@@ -483,7 +465,7 @@ func (s *SQLiteStore) GetImages() ([]*Image, error) {
 	return images, nil
 }
 
-func (s *SQLiteStore) NewImageForUsername(username string) string {
+func (s *PostgresStore) NewImageForUsername(username string) string {
 	images, err := s.GetImages()
 	if err != nil {
 		return err.Error()
@@ -494,7 +476,7 @@ func (s *SQLiteStore) NewImageForUsername(username string) string {
 	return image.Name
 }
 
-func (s *SQLiteStore) GetPlayerForAccount(username string) (*Player, error) {
+func (s *PostgresStore) GetPlayerForAccount(username string) (*Player, error) {
 	acc, err := s.GetAccountByUsername(username)
 	if err != nil {
 		return nil, err
@@ -502,8 +484,8 @@ func (s *SQLiteStore) GetPlayerForAccount(username string) (*Player, error) {
 	return NewPlayer(username, "", acc.ImageName, false, true, acc.NewWordCount, acc.WordCount), nil
 }
 
-func (s *SQLiteStore) GetOwners() ([]*Player, error) {
-	rows, err := s.db.Query("select * from player where is_owner = ?", true)
+func (s *PostgresStore) GetOwners() ([]*Player, error) {
+	rows, err := s.db.Query("select * from player where is_owner = $1", true)
 	if err != nil {
 		return nil, err
 	}
@@ -519,8 +501,8 @@ func (s *SQLiteStore) GetOwners() ([]*Player, error) {
 	return owners, nil
 }
 
-func (s *SQLiteStore) GetLobbyForOwner(owner string) (string, error) {
-	rows, err := s.db.Query("select lobby_code from player where name = ? and is_owner = ?", owner, true)
+func (s *PostgresStore) GetLobbyForOwner(owner string) (string, error) {
+	rows, err := s.db.Query("select lobby_code from player where name = $1 and is_owner = $2", owner, true)
 	if err != nil {
 		return "", err
 	}
@@ -528,6 +510,7 @@ func (s *SQLiteStore) GetLobbyForOwner(owner string) (string, error) {
 	var lobbyCode string
 	defer rows.Close()
 	for rows.Next() {
+		defer rows.Close()
 		err := rows.Scan(&lobbyCode)
 		if err != nil {
 			return "", err
@@ -544,15 +527,15 @@ func (s *SQLiteStore) GetLobbyForOwner(owner string) (string, error) {
 	return lobbyCodes[0], nil
 }
 
-func (s *SQLiteStore) DeletePlayersForLobby(lobbyCode string) error {
-	_, err := s.db.Exec("delete from player where lobby_code = ?", lobbyCode)
+func (s *PostgresStore) DeletePlayersForLobby(lobbyCode string) error {
+	_, err := s.db.Exec("delete from player where lobby_code = $1", lobbyCode)
 	err = s.IncrementPlayerCount(lobbyCode, -1)
 	return err
 }
 
-func (s *SQLiteStore) AddPlayerToLobby(lobbyCode string, player *Player) error {
+func (s *PostgresStore) AddPlayerToLobby(lobbyCode string, player *Player) error {
 	_, err := s.db.Exec(
-		"insert into player (name, lobby_code, image_name, is_owner, has_account, target_word, points, new_word_count, word_count) values (?, ?, ?, ?, ?, ?, ?, ?, ?)",
+		"insert into player (name, lobby_code, image_name, is_owner, has_account, target_word, points, new_word_count, word_count) values ($1, $2, $3, $4, $5, $6, $7, $8, $9)",
 		player.Name,
 		lobbyCode,
 		player.ImageName,
@@ -571,12 +554,12 @@ func (s *SQLiteStore) AddPlayerToLobby(lobbyCode string, player *Player) error {
 	return err
 }
 
-func (s *SQLiteStore) IncrementPlayerCount(lobbyCode string, increment int) error {
-	_, err := s.db.Exec("update lobby set player_count = player_count + ? where lobby_code = ?", increment, lobbyCode)
+func (s *PostgresStore) IncrementPlayerCount(lobbyCode string, increment int) error {
+	_, err := s.db.Exec("update lobby set player_count = player_count + $1 where lobby_code = $2", increment, lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) GetLobbies() ([]*Lobby, error) {
+func (s *PostgresStore) GetLobbies() ([]*Lobby, error) {
 	rows, err := s.db.Query("select * from lobby")
 	if err != nil {
 		return nil, err
@@ -593,13 +576,13 @@ func (s *SQLiteStore) GetLobbies() ([]*Lobby, error) {
 	return lobbies, nil
 }
 
-func (s *SQLiteStore) DeleteLobby(lobbyCode string) error {
-	_, err := s.db.Exec("delete from lobby where lobby_code = ?", lobbyCode)
+func (s *PostgresStore) DeleteLobby(lobbyCode string) error {
+	_, err := s.db.Exec("delete from lobby where lobby_code = $1", lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) GetLobbyByCode(lobbyCode string) (*Lobby, error) {
-	rows, err := s.db.Query("select * from lobby where lobby_code = ?", lobbyCode)
+func (s *PostgresStore) GetLobbyByCode(lobbyCode string) (*Lobby, error) {
+	rows, err := s.db.Query("select * from lobby where lobby_code = $1", lobbyCode)
 	if err != nil {
 		return nil, err
 	}
@@ -610,15 +593,15 @@ func (s *SQLiteStore) GetLobbyByCode(lobbyCode string) (*Lobby, error) {
 	return nil, fmt.Errorf("lobby %s not found", lobbyCode)
 }
 
-func (s *SQLiteStore) EditGameMode(lobbyCode string, gameMode c.GameMode) error {
-	_, err := s.db.Exec("update lobby set game_mode = ? where lobby_code = ?", gameMode, lobbyCode)
+func (s *PostgresStore) EditGameMode(lobbyCode string, gameMode c.GameMode) error {
+	_, err := s.db.Exec("update lobby set game_mode = $1 where lobby_code = $2", gameMode, lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) GetCombination(a, b string) (*string, bool, error) {
+func (s *PostgresStore) GetCombination(a, b string) (*string, bool, error) {
 	a, b = u.SortAB(a, b)
 	var result string
-	err := s.db.QueryRow("SELECT result FROM combination WHERE a = ? AND b = ?", a, b).Scan(&result)
+	err := s.db.QueryRow("select result from combination where a = $1 AND b = $2", a, b).Scan(&result)
 	if err == sql.ErrNoRows {
 		return nil, false, nil
 	} else if err != nil {
@@ -627,10 +610,10 @@ func (s *SQLiteStore) GetCombination(a, b string) (*string, bool, error) {
 	return &result, true, nil
 }
 
-func (s *SQLiteStore) AddCombination(combi *Combination) error {
+func (s *PostgresStore) AddCombination(combi *Combination) error {
 	a, b := u.SortAB(combi.A, combi.B)
 	_, err := s.db.Exec(
-		"insert or ignore into combination (a, b, result, depth) values (?, ?, ?, ?)",
+		"insert into combination (a, b, result, depth) values ($1, $2, $3, $4) on conflict do nothing",
 		a,
 		b,
 		combi.Result,
@@ -639,45 +622,61 @@ func (s *SQLiteStore) AddCombination(combi *Combination) error {
 	return err
 }
 
-func (s *SQLiteStore) AddNewCombination(a, b, result string) error {
+func (s *PostgresStore) AddNewCombination(a, b, result string) error {
 	a, b = u.SortAB(a, b)
 	aDepth := 0
 	bDepth := 0
-	err := s.db.QueryRow("select depth from word where word = ?", a).Scan(&aDepth)
+	err := s.db.QueryRow("select depth from word where word = $1", a).Scan(&aDepth)
 	if err != nil {
 		return err
 	}
-	err = s.db.QueryRow("select depth from word where word = ?", b).Scan(&bDepth)
+	err = s.db.QueryRow("select depth from word where word = $1", b).Scan(&bDepth)
 	if err != nil {
 		return err
 	}
 	depth := max(aDepth, bDepth) + 1
 	_, err = s.db.Exec(
-		"insert or ignore into combination (a, b, result, depth) values (?, ?, ?, ?)",
+		"insert into combination (a, b, result, depth) values ($1, $2, $3, $4) on conflict do nothing",
 		a,
 		b,
 		result,
 		depth,
 	)
-	oldReachability := 0.0
-	err = s.db.QueryRow("select reachability from word where word = ?", result).Scan(&oldReachability)
+	updateDepth := depth
+	oldDepth := 999
+	err = s.db.QueryRow("select depth from word where word = $1", result).Scan(&oldDepth)
 	if err != nil && err != sql.ErrNoRows {
+		log.Printf("No depth for word %s", result)
+	}
+	oldReachability := 0.0
+	oldWeight := 0.25
+	newWeight := 0.75
+	err = s.db.QueryRow("select reachability from word where word = $1", result).Scan(&oldReachability)
+	if err != nil && err != sql.ErrNoRows {
+		oldWeight = 0.0
+		newWeight = 1.0
 		log.Printf("No reachability for word %s", result)
 	}
-	reachability := 1.0 / float64(int(1)<<uint(depth))
+	if oldDepth < depth {
+		updateDepth = oldDepth
+		oldWeight = 0.75
+		newWeight = 0.25
+	}
+
+	reachability := newWeight*(1.0/float64(int(1)<<uint(depth))) + oldWeight*oldReachability
 	_, err = s.db.Exec(
-		"insert or ignore into word (word, depth, reachability) values (?, ?, ?)",
+		"insert into word (word, depth, reachability) values ($1, $2, $3) on conflict do nothing",
 		result,
-		depth,
-		reachability+oldReachability,
+		updateDepth,
+		reachability,
 	)
 	return err
 }
 
-func (s *SQLiteStore) AddWord(word *Word) error {
+func (s *PostgresStore) AddWord(word *Word) error {
 	w := strings.ToLower(word.Word)
 	_, err := s.db.Exec(
-		"insert or ignore into word (word, depth, reachability) values (?, ?, ?)",
+		"insert into word (word, depth, reachability) values ($1, $2, $3) on conflict do nothing",
 		w,
 		word.Depth,
 		word.Reachability,
@@ -685,8 +684,8 @@ func (s *SQLiteStore) AddWord(word *Word) error {
 	return err
 }
 
-func (s *SQLiteStore) GetTargetWords(minReachability, maxReachability float64, maxDepth int) ([]string, error) {
-	rows, err := s.db.Query("select * from word where reachability >= ? and reachability <= ? and depth <= ?", minReachability, maxReachability, maxDepth)
+func (s *PostgresStore) GetTargetWords(minReachability, maxReachability float64, maxDepth int) ([]string, error) {
+	rows, err := s.db.Query("select * from word where reachability >= $1 and reachability <= $2 and depth <= $3", minReachability, maxReachability, maxDepth)
 	if err != nil {
 		return nil, err
 	}
@@ -703,7 +702,7 @@ func (s *SQLiteStore) GetTargetWords(minReachability, maxReachability float64, m
 	return targetWords, nil
 }
 
-func (s *SQLiteStore) GetTargetWord(minReachability, maxReachability float64, maxDepth int) (string, error) {
+func (s *PostgresStore) GetTargetWord(minReachability, maxReachability float64, maxDepth int) (string, error) {
 	targetWords, err := s.GetTargetWords(minReachability, maxReachability, maxDepth)
 	if err != nil {
 		return "", err
@@ -711,9 +710,9 @@ func (s *SQLiteStore) GetTargetWord(minReachability, maxReachability float64, ma
 	return targetWords[rand.Intn(len(targetWords))], nil
 }
 
-func (s *SQLiteStore) AddPlayerWord(playerName, word, lobbyCode string) error {
+func (s *PostgresStore) AddPlayerWord(playerName, word, lobbyCode string) error {
 	_, err := s.db.Exec(
-		"insert or ignore into player_word (player_name, word, lobby_code) values (?, ?, ?)",
+		"insert into player_word (player_name, word, lobby_code) values ($1, $2, $3) on conflict do nothing",
 		playerName,
 		word,
 		lobbyCode,
@@ -721,18 +720,18 @@ func (s *SQLiteStore) AddPlayerWord(playerName, word, lobbyCode string) error {
 	return err
 }
 
-func (s *SQLiteStore) IsPlayerWord(playerName, word, lobbyCode string) (bool, error) {
+func (s *PostgresStore) IsPlayerWord(playerName, word, lobbyCode string) (bool, error) {
 	var count int
-	err := s.db.QueryRow("select count(*) from player_word where player_name = ? and word = ? and lobby_code = ?", playerName, word, lobbyCode).Scan(&count)
+	err := s.db.QueryRow("select count(*) from player_word where player_name = $1 and word = $2 and lobby_code = $3", playerName, word, lobbyCode).Scan(&count)
 	if err != nil {
 		return false, err
 	}
 	return count > 0, nil
 }
 
-func (s *SQLiteStore) SetPlayerTargetWord(playerName, targetWord, lobbyCode string) error {
+func (s *PostgresStore) SetPlayerTargetWord(playerName, targetWord, lobbyCode string) error {
 	_, err := s.db.Exec(
-		"update player set target_word = ? where name = ? and lobby_code = ?",
+		"update player set target_word = $1 where name = $2 and lobby_code = $3",
 		targetWord,
 		playerName,
 		lobbyCode,
@@ -740,17 +739,17 @@ func (s *SQLiteStore) SetPlayerTargetWord(playerName, targetWord, lobbyCode stri
 	return err
 }
 
-func (s *SQLiteStore) GetPlayerTargetWord(playerName, lobbyCode string) (string, error) {
+func (s *PostgresStore) GetPlayerTargetWord(playerName, lobbyCode string) (string, error) {
 	var targetWord string
-	err := s.db.QueryRow("select target_word from player where name = ? and lobby_code = ?", playerName, lobbyCode).Scan(&targetWord)
+	err := s.db.QueryRow("select target_word from player where name = $1 and lobby_code = $2", playerName, lobbyCode).Scan(&targetWord)
 	if err != nil {
 		return "", err
 	}
 	return targetWord, nil
 }
 
-func (s *SQLiteStore) GetPlayerWords(playerName, lobbyCode string) ([]string, error) {
-	rows, err := s.db.Query("select * from player_word where player_name = ? and lobby_code = ? order by timestamp asc", playerName, lobbyCode)
+func (s *PostgresStore) GetPlayerWords(playerName, lobbyCode string) ([]string, error) {
+	rows, err := s.db.Query("select * from player_word where player_name = $1 and lobby_code = $2 order by timestamp asc", playerName, lobbyCode)
 	if err != nil {
 		return nil, err
 	}
@@ -766,21 +765,21 @@ func (s *SQLiteStore) GetPlayerWords(playerName, lobbyCode string) ([]string, er
 	return words, nil
 }
 
-func (s *SQLiteStore) DeletePlayerWordsByLobbyCode(lobbyCode string) error {
-	_, err := s.db.Exec("delete from player_word where lobby_code = ?", lobbyCode)
+func (s *PostgresStore) DeletePlayerWordsByLobbyCode(lobbyCode string) error {
+	_, err := s.db.Exec("delete from player_word where lobby_code = $1", lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) DeletePlayerWordsByPlayerAndLobbyCode(playerName, lobbyCode string) error {
-	_, err := s.db.Exec("delete from player_word where player_name = ? and lobby_code = ?", playerName, lobbyCode)
+func (s *PostgresStore) DeletePlayerWordsByPlayerAndLobbyCode(playerName, lobbyCode string) error {
+	_, err := s.db.Exec("delete from player_word where player_name = $1 and lobby_code = $2", playerName, lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) GetWordCountByLobbyCode(lobbyCode string) ([]*dto.PlayerWordCount, error) {
+func (s *PostgresStore) GetWordCountByLobbyCode(lobbyCode string) ([]*dto.PlayerWordCount, error) {
 	query := `
 	select player_name, COUNT(*) as word_count
 	from player_word
-	where lobby_code = ?
+	where lobby_code = $1
 	group by player_name
 	order by word_count desc
 	`
@@ -801,8 +800,8 @@ func (s *SQLiteStore) GetWordCountByLobbyCode(lobbyCode string) ([]*dto.PlayerWo
 	return wordCounts, nil
 }
 
-func (s *SQLiteStore) GetPlayersWithAccount(lobbyCode string) ([]*Account, error) {
-	rows, err := s.db.Query("select * from player where lobby_code = ? and has_account = ?", lobbyCode, true)
+func (s *PostgresStore) GetPlayersWithAccount(lobbyCode string) ([]*Account, error) {
+	rows, err := s.db.Query("select * from player where lobby_code = $1 and has_account = $2", lobbyCode, true)
 	if err != nil {
 		return nil, err
 	}
@@ -822,7 +821,7 @@ func (s *SQLiteStore) GetPlayersWithAccount(lobbyCode string) ([]*Account, error
 	return accounts, nil
 }
 
-func (s *SQLiteStore) UpdateAccountWinsAndLosses(lobbyCode, winner string) error {
+func (s *PostgresStore) UpdateAccountWinsAndLosses(lobbyCode, winner string) error {
 	accounts, err := s.GetPlayersWithAccount(lobbyCode)
 	if err != nil {
 		return err
@@ -840,36 +839,36 @@ func (s *SQLiteStore) UpdateAccountWinsAndLosses(lobbyCode, winner string) error
 	return nil
 }
 
-func (s *SQLiteStore) IncrementPlayerPoints(playerName, lobbyCode string, points int) error {
-	_, err := s.db.Exec("update player set points = points + ? where name = ? and lobby_code = ?", points, playerName, lobbyCode)
+func (s *PostgresStore) IncrementPlayerPoints(playerName, lobbyCode string, points int) error {
+	_, err := s.db.Exec("update player set points = points + $1 where name = $2 and lobby_code = $3", points, playerName, lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) ResetPlayerPoints(lobbyCode string) error {
-	_, err := s.db.Exec("update player set points = 0 where lobby_code = ?", lobbyCode)
+func (s *PostgresStore) ResetPlayerPoints(lobbyCode string) error {
+	_, err := s.db.Exec("update player set points = 0 where lobby_code = $1", lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) SetIsOwner(username string, setOwner bool) error {
+func (s *PostgresStore) SetIsOwner(username string, setOwner bool) error {
 	if !setOwner {
-		_, err := s.db.Exec("update account set is_owner = ? where username = ?", setOwner, username)
+		_, err := s.db.Exec("update account set is_owner = $1 where username = $2", setOwner, username)
 		return err
 	}
 
 	var isOwner bool
-	err := s.db.QueryRow("select is_owner from account where username = ?", username).Scan(&isOwner)
+	err := s.db.QueryRow("select is_owner from account where username = $1", username).Scan(&isOwner)
 	if err != nil {
 		return err
 	}
 	if isOwner {
-		return fmt.Errorf("User is already owner!")
+		return fmt.Errorf("user is already owner!")
 	}
-	_, err = s.db.Exec("update account set is_owner = ? where username = ?", setOwner, username)
+	_, err = s.db.Exec("update account set is_owner = $1 where username = $2", setOwner, username)
 	return err
 }
 
-func (s *SQLiteStore) SelectWinnerByPoints(lobbyCode string) (string, error) {
-	rows, err := s.db.Query("select name from player where lobby_code = ? order by points desc", lobbyCode)
+func (s *PostgresStore) SelectWinnerByPoints(lobbyCode string) (string, error) {
+	rows, err := s.db.Query("select name from player where lobby_code = $1 order by points desc", lobbyCode)
 	if err != nil {
 		return "", err
 	}
@@ -889,14 +888,14 @@ func (s *SQLiteStore) SelectWinnerByPoints(lobbyCode string) (string, error) {
 	return winners[0], nil
 }
 
-func (s *SQLiteStore) DeleteAccount(username string) error {
-	_, err := s.db.Exec("delete from account where username = ?", username)
+func (s *PostgresStore) DeleteAccount(username string) error {
+	_, err := s.db.Exec("delete from account where username = $1", username)
 	return err
 }
 
-func (s *SQLiteStore) GetChallengeEntries() ([]*Challenger, error) {
+func (s *PostgresStore) GetChallengeEntries() ([]*Challenger, error) {
 	today := time.Now().Format("2006-01-02")
-	rows, err := s.db.Query("select * from daily_challenge where timestamp = ?", today)
+	rows, err := s.db.Query("select * from daily_challenge where timestamp = $1", today)
 	if err != nil {
 		return nil, err
 	}
@@ -912,52 +911,16 @@ func (s *SQLiteStore) GetChallengeEntries() ([]*Challenger, error) {
 	return entries, nil
 }
 
-func (s *SQLiteStore) GetImageByUsername(username string) ([]byte, error) {
+func (s *PostgresStore) GetImageByUsername(username string) ([]byte, error) {
 	var imageName string
-	err := s.db.QueryRow("select image_name from account where username = ?", username).Scan(&imageName)
+	err := s.db.QueryRow("select image_name from account where username = $1", username).Scan(&imageName)
 	if err != nil {
 		return nil, err
 	}
 	return s.GetImage(imageName)
 }
 
-func (s *SQLiteStore) CreateSession(session *Session) error {
-	query := `insert into session 
-	(id, refresh_token, username, is_revoked, created_at, expires_at)
-	values (?, ?, ?, ?, ?, ?)`
-	_, err := s.db.Exec(
-		query,
-		session.ID,
-		session.RefreshToken,
-		session.Username,
-		session.IsRevoked,
-		session.CreatedAt,
-		session.ExpiresAt,
-	)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (s *SQLiteStore) GetSession(id string) (*Session, error) {
-	rows, err := s.db.Query("select * from session where id = ?", id)
-	if err != nil {
-		return nil, err
-	}
-	for rows.Next() {
-		defer rows.Close()
-		return scanIntoSession(rows)
-	}
-	return nil, fmt.Errorf("session %s not found", id)
-}
-
-func (s *SQLiteStore) RevokeSession(id string) error {
-	_, err := s.db.Exec("update session set is_revoked = ? where id = ?", true, id)
-	return err
-}
-
-func (s *SQLiteStore) GetAchievements() ([]*AchievementEntry, error) {
+func (s *PostgresStore) GetAchievements() ([]*AchievementEntry, error) {
 	rows, err := s.db.Query("select * from achievement")
 	if err != nil {
 		return nil, err
@@ -974,27 +937,27 @@ func (s *SQLiteStore) GetAchievements() ([]*AchievementEntry, error) {
 	return entries, nil
 }
 
-func (s *SQLiteStore) UpdateAccountWordCount(username string, newWordCount, wordCount int) error {
-	_, err := s.db.Exec("update account set new_word_count = ?, word_count = ? where username = ?", newWordCount, wordCount, username)
+func (s *PostgresStore) UpdateAccountWordCount(username string, newWordCount, wordCount int) error {
+	_, err := s.db.Exec("update account set new_word_count = $1, word_count = $2 where username = $3", newWordCount, wordCount, username)
 	return err
 }
 
-func (s *SQLiteStore) UpdatePlayerWordCount(playerName, lobbyCode string, newWordCount, wordCount int) error {
-	_, err := s.db.Exec("update player set new_word_count = ?, word_count = ? where name = ? and lobby_code = ?", newWordCount, wordCount, playerName, lobbyCode)
+func (s *PostgresStore) UpdatePlayerWordCount(playerName, lobbyCode string, newWordCount, wordCount int) error {
+	_, err := s.db.Exec("update player set new_word_count = $1, word_count = $2 where name = $3 and lobby_code = $4", newWordCount, wordCount, playerName, lobbyCode)
 	return err
 }
 
-func (s *SQLiteStore) AddAchievementImage(data []byte, name string) error {
+func (s *PostgresStore) AddAchievementImage(data []byte, name string) error {
 	_, err := s.db.Exec(
-		"insert or replace into achievement_image (name, data) values (?, ?)",
+		"insert into achievement_image (name, data) values ($1, $2) on conflict (name) do update set data = $2",
 		name,
 		data,
 	)
 	return err
 }
 
-func (s *SQLiteStore) GetAchievementImage(name string) ([]byte, error) {
-	rows, err := s.db.Query("select * from achievement_image where name = ?", name)
+func (s *PostgresStore) GetAchievementImage(name string) ([]byte, error) {
+	rows, err := s.db.Query("select * from achievement_image where name = $1", name)
 	if err != nil {
 		return nil, err
 	}
@@ -1016,8 +979,8 @@ func (s *SQLiteStore) GetAchievementImage(name string) ([]byte, error) {
 	return images[0].Data, nil
 }
 
-func (s *SQLiteStore) GetAchievementByTitle(title string) (*AchievementEntry, error) {
-	rows, err := s.db.Query("select * from achievement where title = ?", title)
+func (s *PostgresStore) GetAchievementByTitle(title string) (*AchievementEntry, error) {
+	rows, err := s.db.Query("select * from achievement where title = $1", title)
 	if err != nil {
 		return nil, err
 	}
@@ -1028,8 +991,8 @@ func (s *SQLiteStore) GetAchievementByTitle(title string) (*AchievementEntry, er
 	return nil, fmt.Errorf("Achievement %s not found", title)
 }
 
-func (s *SQLiteStore) GetAchievementsForUser(username string) ([]string, error) {
-	rows, err := s.db.Query("select achievement_title from unlocked where username = ?", username)
+func (s *PostgresStore) GetAchievementsForUser(username string) ([]string, error) {
+	rows, err := s.db.Query("select achievement_title from unlocked where username = $1", username)
 	if err != nil {
 		return nil, err
 	}
